Manager class removal and Employee class updating

+++ src/main/java/com/gokhan/engcompany/service/HeadDepartmentService.java
@@ -1,10 +1,13 @@
 package com.gokhan.engcompany.service;
 
+import com.gokhan.engcompany.dto.DepartmentDto;
 import com.gokhan.engcompany.dto.HeadDepartmentDto;
 import com.gokhan.engcompany.entity.Department;
+import com.gokhan.engcompany.entity.Employee;
 import com.gokhan.engcompany.entity.HeadDepartment;
 import com.gokhan.engcompany.enums.DepartmentType;
 import com.gokhan.engcompany.repository.HeadDepartmentRepository;
+import com.gokhan.engcompany.request.HeadDepartmentRequest;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -17,31 +20,37 @@ public class HeadDepartmentService {
     HeadDepartmentRepository repository;
 
     @Autowired
-    ManagerService managerService;
+    EmployeeService employeeService;
 
     @Autowired
     DepartmentService departmentService;
 
 
-    public HeadDepartmentDto createHeadDepartment(DepartmentType departmentType) {
-        checkIfHeadDepartmentExists(departmentType);
+    public HeadDepartmentDto createHeadDepartment(HeadDepartmentRequest headDepartmentRequest) {
+        checkIfHeadDepartmentExists(headDepartmentRequest.departmentType);
         HeadDepartment headDepartment = new HeadDepartment();
-        headDepartment.setDepartmentType(departmentType);
+        headDepartment.setDepartmentType(headDepartmentRequest.departmentType);
         repository.save(headDepartment);
         return headDepartment.toDto();
     }
 
 
-    public HeadDepartmentDto updateManager(int managerId, int headDepartmentId) {
+    public HeadDepartmentDto updateManager(int employeeId, int headDepartmentId) {
 
         HeadDepartment headDepartment = repository.findById(headDepartmentId).get();
-        headDepartment.setManager(managerService.getManagerEntity(managerId));
+        Employee manager = employeeService.getManagerIfManager(employeeId);
+        if (manager != null) {
+            headDepartment.setManager(manager);
+            manager.setManagerOf(headDepartmentId);
+        } else { // custom throw add
+            return null;
+        }
         return repository.save(headDepartment).toDto();
     }
 
     public HeadDepartmentDto addDepartment(int departmentId, int headDepartmentId) {
 
-        if (repository.existsByHeadDepartmentId(headDepartmentId)) {
+        if (repository.existsByHeadDepartmentId(headDepartmentId).get()) {
             HeadDepartment headDepartment = repository.findById(headDepartmentId).get();
             return (repository.save(checkIfDepartmentNotAPartOfADepartmentList(headDepartment,departmentId))
                     .toDto());
@@ -51,8 +60,10 @@ public class HeadDepartmentService {
 
     }
 
-    public HeadDepartment checkIfDepartmentNotAPartOfADepartmentList(HeadDepartment headDepartment, int departmentId) {
-        if (headDepartment.getDepartmentList().stream().map(Department::getDepartmentId).equals(departmentId)) {
+    public HeadDepartment checkIfDepartmentNotAPartOfADepartmentList(HeadDepartment headDepartment,
+                                                                     int departmentId) {
+        if (headDepartment.getDepartmentList().stream().
+                map(Department::getDepartmentId).equals(departmentId)) {
             throw new EntityExistsException();
         } else {
             Department department = departmentService.getDepartmentEntity(departmentId);
@@ -69,17 +80,29 @@ public class HeadDepartmentService {
         }
     }
 
-    public HeadDepartmentDto getHeadDepartment(int headDepartmentId) {
+    public HeadDepartment getHeadDepartmentEntity(int headDepartmentId) {
 
         HeadDepartment headDepartment = repository.findById(headDepartmentId).orElse(null);
-        return headDepartment.toDto();
+        return headDepartment;
     }
 
     public String deleteHeadDepartment(int headDepartmentId) {
 
-        repository.deleteById(headDepartmentId);
-        return "Bölüm silindi";
+        if (getHeadDepartmentEntity(headDepartmentId) != null) {
+            repository.deleteById(headDepartmentId);
+            return "Bölüm silindi";
+        } else {
+            return "Bölüm silinemedi, çünkü yok";
+        }
     }
 
+    public HeadDepartmentDto getHeadDepartmentDto(int headDepartmentId) {
 
+        HeadDepartment headDepartment = getHeadDepartmentEntity(headDepartmentId);
+        if ( headDepartment == null) {
+            return new HeadDepartmentDto("This Employee is not exist");
+        } else {
+            return headDepartment.toDto();
+        }
+    }
 }
